renamed file to class_home_applianc.py
