+++ src/akkudoktoreos/class_home_appliance.py
