+++ src/akkudoktoreos/class_optimize.py
@@ -6,7 +6,7 @@ from deap import algorithms, base, creator, tools
 
 from akkudoktoreos.class_akku import PVAkku
 from akkudoktoreos.class_ems import EnergieManagementSystem
-from akkudoktoreos.class_haushaltsgeraet import HomeAppliance
+from akkudoktoreos.class_home_appliance import HomeAppliance
 from akkudoktoreos.class_inverter import Wechselrichter
 from akkudoktoreos.config import AppConfig
 from akkudoktoreos.visualize import visualisiere_ergebnisse
